rename cartodb -> carto. closes #187
